Code/requirements.txt
aioconsole==0.6.1
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyascii==0.3.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
confection==0.0.4
contractions==0.1.73
cymem==2.0.7
datasets==2.11.0
decorator==5.1.1
dill==0.3.6
Django==4.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl#sha256=b492fd2dfeee931dd60232fd1a122bb5c3542ebcd5d614f11844647050aa539e
entrypoints==0.4
evaluate==0.4.0
filelock==3.12.0
flatbuffers==23.5.8
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.4
humanfriendly==10.0
idna==3.4
image==1.5.33
importlib-metadata==6.6.0
install==1.3.5
Jinja2==3.1.2
jiwer==3.0.1
joblib==1.2.0
jsonschema==4.17.3
langcodes==3.3.0
load-dotenv==0.1.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.24.3
onnx==1.14.0
onnxruntime==1.14.1
openai==0.27.6
openai-secret-manager==0.0.2
packaging==23.1
pandas==2.0.0
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
protobuf==3.20.3
pyahocorasick==2.0.0
pyarrow==11.0.0
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rapidfuzz==2.13.7
regex==2023.3.23
requests==2.28.2
responses==0.18.0
rich==13.3.5
scikit-learn==1.2.2
scipy==1.10.1
semantic-kernel==0.2.6.dev0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==2.4.6
streamlit==1.22.0
streamlit-chat==0.0.2.2
sympy==1.11.1
tenacity==8.2.2
textsearch==0.0.24
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
Top==1.0.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
transformers==4.28.1
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uuid==1.30
validators==0.20.0
wasabi==1.1.1
watchdog==3.0.0
xxhash==3.2.0
yarl==1.9.1
zipp==3.15.0
